fix: (virtual-background) remove selected background

+++ react/features/virtual-background/components/VirtualBackgroundDialog.js
@@ -47,7 +47,7 @@ function VirtualBackground({ _apiBase, _trackExist, _virtualSource, dispatch, t 
             try {
                 const resp = await axios.get(`${_apiBase}/backgrounds?pagination=false`);
                 setData(resp.data);
-                setImages([...images, ...resp.data.docs]);
+                setImages([...resp.data.docs, ...images]);
                 isloading(false);
             } catch {
                 isloading(false);
@@ -65,6 +65,12 @@ function VirtualBackground({ _apiBase, _trackExist, _virtualSource, dispatch, t 
     };
 
     const removeBackground = async image => {
+        if (image._id === selected) {
+            const found = images.findIndex(item => item === image);
+            const next = found > 0 ? images[found-1] : images[0];
+    
+            setImageBackground(next);
+        }
         setImages(images.filter(item => image !== item));
         axios.delete(`${_apiBase}/backgrounds/${image._id}`);
     };
